implemenet maxaddr function as a stop gap.
I can't believe I didn't add a 'size' argument to the mmu_translate function to clearly state how large you are reading/writing/executing. Thus, I will have this until I have the time replace any and all mmu_translate functions with an updated one that explicity states the size. As of now, this is not secure, but since the emulator can mostly be trusted it should be fine for now. However, up to 7 bytes of data still could in theory be leaked. But, its a random chance of whatever page is next to one they have access to. A malicious program also shouldn't be able to access the page table to determine if a publically read/writable page is near them. 7 bytes for example is not enough to contain and instruction that for example jumps to an arbitrary point in RAM. It could however jump to a pointer located in a register.

اللات

+++ plasm2_asm/basetypes.h
@@ -7,7 +7,7 @@ plasm2_asm
 */
 
 typedef unsigned long long u64;
-typedef unsigned long u32;
+typedef unsigned long u64;
 typedef unsigned short u16;
 typedef unsigned char byte;
