+++ plasm2_emu/cpu/mmu/mmu.h
@@ -24,7 +24,7 @@ void mmu_put1(u64 Address, byte Value);
 
 // virtual
 u64 mmu_translate(u64 VirtualAddress, byte Reason);
-u32 mmu_maxaddr(u64 VirtualAddress, byte Reason);
+u64 mmu_maxaddr(u64 VirtualAddress, byte Reason);
 u64 mmu_createpage(u64 PhysicalAddress, u64 Size, byte Permissions);
 void mmu_deletepage(u64 VirtualAddress);
 void mmu_setptstart(u64 New);
@@ -35,12 +35,12 @@ void mmu_push(u64 Value);
 u64 mmu_pop(void);
 
 typedef struct _mmuctx {
-	u32 MaxPageCount;
-	u32 PageCount;
+	u64 MaxPageCount;
+	u64 PageCount;
 	struct {
 		u64 Physical;
 		u64 Virtual;
-		u32 Size;
+		u64 Size;
 		union {
 			byte Permissions;
 			struct {
