+++ plasm2_emu/devices/devices.h
@@ -12,14 +12,14 @@ void devices_init(void);
 void devices_shutdown(void);
 void devices_collect(void);
 
-u64 devicesi_statusquery(u32 Device);
-void devicesi_sendcommand(u32 Device, u64 Command);
-void devicesi_senddata(u32 Device, u64 Command);
-u64  devicesi_getdata(u32 Device);
-void devicesi_reset(u32 Device);
-void devicesi_off(u32 Device);
-void devicesi_on(u32 Device);
-u32  devicesi_devcount(void);
+u64 devicesi_statusquery(u64 Device);
+void devicesi_sendcommand(u64 Device, u64 Command);
+void devicesi_senddata(u64 Device, u64 Command);
+u64  devicesi_getdata(u64 Device);
+void devicesi_reset(u64 Device);
+void devicesi_off(u64 Device);
+void devicesi_on(u64 Device);
+u64  devicesi_devcount(void);
 
 #define DEVTYPE_KB			0x10000001
 #define DEVTYPE_MOUSE		0x10000002
@@ -40,17 +40,17 @@ u32  devicesi_devcount(void);
 #define DEVSTATUS_RSN_CLEAR  0x00FF0000
 
 typedef struct _devicesctx {
-	u32 DeviceCount;
+	u64 DeviceCount;
 	struct {
 		int DeviceType;
 		char DeviceName[32];
 		char DeviceVendor[32];
-		u32 VendorId;
-		u32 DeviceModel;
-		u32 DeviceSerial;
+		u64 VendorId;
+		u64 DeviceModel;
+		u64 DeviceSerial;
 		u64(*Callbacks[7])(u32 DeviceId, u64 Argument);
 		union {
-			u32 FlagsRaw;
+			u64 FlagsRaw;
 			struct {
 				byte Active : 1;
 				byte On : 1;
