+++ plasm2_emu/devices/video/video.h
@@ -12,13 +12,13 @@ void video_shutdown(void);
 void video_collect(void);
 void video_clock(void);
 
-u64  video_statusquery(u32 Device);
-void video_sendcommand(u32 Device, u64 Command);
-void video_senddata(u32 Device, u64 Command);
-u64  video_getdata(u32 Device);
-void video_reset(u32 Device);
-void video_off(u32 Device);
-void video_on(u32 Device);
+u64  video_statusquery(u64 Device);
+void video_sendcommand(u64 Device, u64 Command);
+void video_senddata(u64 Device, u64 Command);
+u64  video_getdata(u64 Device);
+void video_reset(u64 Device);
+void video_off(u64 Device);
+void video_on(u64 Device);
 
 /*
 defaults:
@@ -43,11 +43,11 @@ SuggestWh     07 : Suggest a width (hi32) and height (lo32)
 
 u64 videoi_gettextbuffer(void); // deprecated
 void videoi_settextbuffer(u64 NewOffset); // deprecated
-void videoi_drawline(u16 x1, u16 y1, u16 x2, u16 y2, u32 color); // color pulled from stack
-void videoi_drawrect(u16 x, u16 y, u16 w, u16 h, u32 color);     // ^
-void videoi_drawfill(u16 x, u16 y, u16 w, u16 h, u32 color);     // ^
+void videoi_drawline(u16 x1, u16 y1, u16 x2, u16 y2, u64 color); // color pulled from stack
+void videoi_drawrect(u16 x, u16 y, u16 w, u16 h, u64 color);     // ^
+void videoi_drawfill(u16 x, u16 y, u16 w, u16 h, u64 color);     // ^
 void videoi_copyrect(u16 x, u16 y, u16 w, u16 h, u64 ptr); // ptr pulled from stack. size = (w * h * 4)
-u32  videoi_getwh(void);
+u64  videoi_getwh(void);
 void videoi_suggestwh(u16 w, u16 h);
 
 typedef struct _videoctx {
